requirements.txt
flask~=3.0.2
psycopg2-binary

